lift mode syncing hook logic to context + remove hook
